Remove monkey patches for valid_value checks
