+++ lib/sequent/core/helpers/date_time_validator.rb
@@ -3,15 +3,6 @@ require 'active_model'
 module Sequent
   module Core
     module Helpers
-      # Validates Dates
-      # Automatically included when using a
-      #
-      #   attrs value: Date
-      class DateValidator < ActiveModel::EachValidator
-        def validate_each(subject, attribute, value)
-          subject.errors.add attribute, :invalid_date unless Date.valid_value?(value)
-        end
-      end
       # Validates DateTimes
       # Automatically included when using a
       #
